agents/vertex-adk/pyproject.toml
[project]
name = "vertex-adk-agent"
version = "0.1.0"
description = "Vertex ADK RAG agent for Gmail draft generation"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.5",
    "uvicorn[standard]==0.32.0",
    "google-cloud-aiplatform==1.67.0",
    "google-cloud-secret-manager==2.21.1",
    "google-auth==2.34.1",
    "vertexai==1.67.0",
    "pydantic==2.9.2",
    "python-dotenv==1.0.1"
]

[tool.uv]
package = true

[tool.uv.sources]
vertex-adk-agent = {path = "src"}



